Revert "Minor problems detected during compilation - fix"
This reverts commit bbf7a0bd53ff14bcb94fbc15415a073c9906eb59.

+++ tests/tests/operation_time_tests.cpp
@@ -2987,7 +2987,7 @@ BOOST_AUTO_TEST_CASE( generate_account_subsidies )
       };
 
       const witness_schedule_object& wso = db->get_witness_schedule_object();
-      BOOST_CHECK_EQUAL( wso.account_subsidy_rd.resource_unit, static_cast< size_t >( STEEM_ACCOUNT_SUBSIDY_PRECISION ) );
+      BOOST_CHECK_EQUAL( wso.account_subsidy_rd.resource_unit, STEEM_ACCOUNT_SUBSIDY_PRECISION );
       BOOST_CHECK_EQUAL( wso.account_subsidy_rd.budget_per_time_unit, 5123 );
       BOOST_CHECK(  is_pool_in_equilibrium( int64_t( wso.account_subsidy_rd.pool_eq )  , wso.account_subsidy_rd.budget_per_time_unit, wso.account_subsidy_rd.decay_params ) );
       BOOST_CHECK( !is_pool_in_equilibrium( int64_t( wso.account_subsidy_rd.pool_eq )-1, wso.account_subsidy_rd.budget_per_time_unit, wso.account_subsidy_rd.decay_params ) );
@@ -3065,9 +3065,9 @@ BOOST_AUTO_TEST_CASE( account_subsidy_witness_limits )
          generate_block();
 
          // The transaction fails in generate_block(), meaning it is removed from the local node's transaction list
-         BOOST_CHECK_EQUAL( db->fetch_block_by_number( db->head_block_num() )->transactions.size(), static_cast< size_t >( 0 ) );
+         BOOST_CHECK_EQUAL( db->fetch_block_by_number( db->head_block_num() )->transactions.size(), 0 );
          BOOST_CHECK( db->get_account( "alice" ).pending_claimed_accounts == 0 );
-         BOOST_CHECK_EQUAL( db->_pending_tx.size(), static_cast< size_t >( 0 ) );
+         BOOST_CHECK_EQUAL( db->_pending_tx.size(), 0 );
       } while( db->get< witness_object, by_name >( db->get_scheduled_witness( 1 ) ).schedule == witness_object::timeshare );
 
       db->push_transaction( tx, 0 );
@@ -3075,9 +3075,9 @@ BOOST_AUTO_TEST_CASE( account_subsidy_witness_limits )
       BOOST_CHECK( db->_pending_tx.size() == 1 );
       // But generate another block, as a non-time-share witness, and it works
       generate_block();
-      BOOST_CHECK_EQUAL( db->fetch_block_by_number( db->head_block_num() )->transactions.size(), static_cast< size_t >( 1 ) );
+      BOOST_CHECK_EQUAL( db->fetch_block_by_number( db->head_block_num() )->transactions.size(), 1 );
       BOOST_CHECK( db->get_account( "alice" ).pending_claimed_accounts == 1 );
-      BOOST_CHECK_EQUAL( db->_pending_tx.size(), static_cast< size_t >( 0 ) );
+      BOOST_CHECK_EQUAL( db->_pending_tx.size(), 0 );
 
       while( db->get< witness_object, by_name >( db->get_scheduled_witness( 1 ) ).schedule == witness_object::timeshare )
       {
@@ -3101,7 +3101,7 @@ BOOST_AUTO_TEST_CASE( account_subsidy_witness_limits )
       BOOST_CHECK_EQUAL( db->_pending_tx.size(), n+1 );
       generate_block();
       BOOST_CHECK_EQUAL( db->fetch_block_by_number( db->head_block_num() )->transactions.size(), n );
-      BOOST_CHECK_EQUAL( db->_pending_tx.size(), static_cast< size_t >( 1 ) );
+      BOOST_CHECK_EQUAL( db->_pending_tx.size(), 1 );
    }
    FC_LOG_AND_RETHROW()
 }
